[FEATURE] Option to build & install QCH file with the public API dox
Using the extra-cmake-modules module ECMAddQch this adds the option
to automatically build and install a file in QCH format with the docs
about the public API, which then can be used e.g. in Qt Assistant,
Qt Creator or KDevelop.

Additionally the installed cmake config files will be extended
with a target KF5DAV_QCH containing information about how to "link"
into the generated QCH file, which then can be used in the cmake build
system of other libraries building on this library, by
simply listing this target in "LINK_QCHS" of their ecm_add_qch() usage.
And a respective doxygen tag file with all the metadata about the
generated QCH file and used for the "linking" will be created and
installed.

Pass -DBUILD_QCH=ON to cmake to enable this.