Also use relative hrefs with for carddav-multiget
This applies the same change to carddav as already done for caldav
(Commit bf6628ff1da1661a6485a2199d2222611e1b2ae5 "Replace absolute
URIs with relative ones in calendar-multiget request").

The rationale is some servers apparently having problems when the
DAV:href elements contains non-relative URLs (i.e. including scheme
and host), and return empty result in this case. (One such server
is Telekom Magentacloud/Spica, which is fixed with this change.)

It would seem safer to use the same notation as originally reported
in the DAV collection, but even that does not seem to always work - see
commit abfdd65f0c85332e1c16e5c01c74478d392364ba ("Use URLs in multiget
requests as returned by the Server"), which was reverted as it broke
with KDABs servers 7 years ago.

As nobody has reported any problems with the caldav change above, it is
probably safe to use relative hrefs now.

CCBUG: 386985
CCBUG: 310210