Don't hide when the principals home sets fetch failed
When fetch the collections, if the principals home sets fetch failed
with a temporary error we had weird logic to attempt to do the
collection fetch directly on the same endpoint.

This is dangerous as this can lead the second fetch to succeed on the
same endpoint... in which case we'd report an empty list of collections.

This could happen randomly on transient connectivity issues or issues
with the server. Clients would then wrongly assume the server has no
collection to report. In the case of the Akonadi DAV resources it'd lead
to deleting all the previously known collections, leading to potential
data losses or at best state loss nuisances (like collection attributes,
changing IDs for collections once they are recreated, etc.).

So change the logic to just bubbling up the error, so that the client
can indeed notice something wrong happened and try again later.